requirements.txt
pandas==1.5.2
spacy==3.4.4
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.4.1/en_core_web_trf-3.4.1-py3-none-any.whl
bert-extractive-summarizer==0.10.1
torch==1.13.1
EasyNMT==2.0.2
protobuf==3.20.3